+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -156,7 +156,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-08T11:57:57.422Z",
+    "2026-04-08T13:40:53.886Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {