+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-08T11:57:57.422Z",
+    "2026-04-08T13:40:53.886Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {